requirements.txt
streamlit
langchain-core
langchain-community
ollama
python-dotenv
requests

